Rename classes to re-organize project.
