+++ Card/src/main/java/com/paypal/android/card/CardClient.kt
@@ -79,7 +79,7 @@ class CardClient internal constructor(
             cardAPI.confirmPaymentSource(orderId, cardRequest.card, cardRequest.threeDSecureRequest)
         if (confirmPaymentSourceResponse.payerActionHref == null) {
             val result = confirmPaymentSourceResponse.let {
-                CardResult(it.orderId, it.status, it.paymentSource)
+                CardResult(it.orderID, it.status, it.paymentSource)
             }
             withContext(Dispatchers.Main) {
                 approveOrderListener?.onApproveOrderSuccess(result)
