+++ Card/src/main/java/com/paypal/android/card/api/ConfirmPaymentSourceResponse.kt
@@ -5,7 +5,7 @@ import com.paypal.android.card.model.PurchaseUnit
 import com.paypal.android.core.OrderStatus
 
 internal data class ConfirmPaymentSourceResponse(
-    val orderId: String,
+    val orderID: String,
     val status: OrderStatus? = null,
     val payerActionHref: String? = null,
     val paymentSource: PaymentSource? = null,
