+++ Core/src/main/java/com/paypal/android/core/API.kt
@@ -1,16 +1,11 @@
 package com.paypal.android.core
 
-import java.net.HttpURLConnection.HTTP_OK
-
 class API internal constructor(
     private val configuration: CoreConfig,
     private val http: Http,
     private val httpRequestFactory: HttpRequestFactory
 ) {
 
-    companion object {
-        val SUCCESSFUL_STATUS_CODES = HTTP_OK..299
-    }
     constructor(configuration: CoreConfig) :
             this(configuration, Http(), HttpRequestFactory())
