+++ Core/src/main/java/com/paypal/android/core/HttpResponse.kt
@@ -1,5 +1,7 @@
 package com.paypal.android.core
 
+import java.net.HttpURLConnection
+
 data class HttpResponse(
     val status: Int,
     val headers: Map<String, String> = emptyMap(),
@@ -10,5 +12,9 @@ data class HttpResponse(
         const val STATUS_UNDETERMINED = -1
         const val STATUS_UNKNOWN_HOST = -2
         const val SERVER_ERROR = -3
+
+        val SUCCESSFUL_STATUS_CODES = HttpURLConnection.HTTP_OK..299
     }
+
+    var isSuccessful: Boolean = status in SUCCESSFUL_STATUS_CODES
 }
